backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
python-jose
passlib
bcrypt==4.0.1
